Only call FF if it exceeds message limit
